pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["leo-guinan <leo@copy.ai>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
Django = "^5.1"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.4"}
tweepy = "^4.10.1"
python-decouple = "^3.6"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.1"
djangorestframework-api-key = "2.*"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
django-celery-beat = "^2.3.0"
dj-database-url = "^1.0.0"
psycopg2-binary = "2.9.7"
gunicorn = "^20.1.0"
flower = "^1.2.0"
redis = "^4.3.4"
whitenoise = {extras = ["brotli"], version = "^6.1.0"}
stripe = "^4.2.0"
sendgrid = "^6.9.7"
feedparser = "^6.0.10"
setuptools = "^65.4.1"
importlib-metadata = "^6.11.0"
typing-extensions = "^4.3.0"
cached-property = "^1.5.2"
python-dateutil = "^2.8.2"
pandas = "^1.5.1"
openai = "^1.1.0"
tokenizers = "^0.14.1"
langchain = "^0.1.14"
numpy = "^1.24.1"
scikit-learn = "^1.2.0"
pinecone-client = "^3.0.2"
aiohttp = "^3.8.1"
replicate = "^0.4.0"
boto3 = "^1.26.62"
pydub = "^0.25.1"
chromadb = "0.5.0"
pytube = {git = "https://github.com/leo-guinan/pytube.git"}
tiktoken = "0.6.0"
uuid = "^1.30"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
notion-client = "^2.0.0"
psycopg = "3.1.10"
pymongo = "^4.4.1"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"
lark = "^1.1.7"
tabulate = "^0.9.0"
unstructured = "^0.10.12"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
langchainhub = "^0.1.13"
simple-salesforce = "^1.12.5"
urllib3 = "1.26.12"
pytz = "2023.3"
websockets = "11.0.3"
humanize = "4.7.0"
pydantic-settings = "^2.1.0"
requests = {extras = ["security"], version = "^2.31.0"}
llama-index = "^0.9.46"
langchain-experimental = "^0.0.56"
langchain-openai = "^0.1.1"
langchain-pinecone = "^0.0.3"
anthropic = "^0.25.6"
langchain-anthropic = "^0.1.11"
pyairtable = "^2.3.3"
langsmith = "^0.1.49"
langchain-mongodb = "^0.1.3"
django-storages = {extras = ["s3"], version = "^1.14.2"}
pymupdf = "^1.24.2"
jupyterlab = "^4.1.8"
nest-asyncio = "^1.6.0"
graphviz = "^0.20.3"
playwright = "^1.43.0"
resend = "^1.1.0"
google-api-python-client = "^2.132.0"
markdown = "^3.6"
pdfkit = "^1.0.0"
langchain-chroma = "^0.1.1"
pillow = "^10.3.0"
ell-ai = "^0.0.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.2.1", "setuptools"]
build-backend = "poetry.core.masonry.api"
